adds click listener on mobile in tooltip component
